Change the client method name from obfs3_flash to obfs3_flashproxy.
"flashproxy" is one of the names recognized by flashproxy-client. The
other is "websocket". Since method names aren't allowed to contain a '|'
(#9580), this at least looks like a mechanical conversion from
obfs3|flashproxy to obfs3_flashproxy.

The transport list that you register with; i.e. the --transport option
to flashproxy-client and friends, is still "obfs3|websocket". That's
because the facilitator actually parses this string, and proxies use the
last part to decide how they will connect to you.